# solution/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления данными клиентов, рекламодателей, рекламными кампаниями, показом объявлений, статистикой и управлением \"текущим днём\" в системе.",
        "title": "PROD Backend 2025 Advertising Platform API",
        "contact": {},
        "license": {
            "name": "GPL 3.0"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/clients/bulk": {
            "post": {
                "description": "Создает новых или обновляет существующих клиентов",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clients"
                ],
                "summary": "Массовое создание/обновление клиентов",
                "parameters": [
                    {
                        "description": "Clients",
                        "name": "clients",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.User"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.User"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/clients/{clientId}": {
            "get": {
                "description": "Возвращает информацию о клиенте по его ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clients"
                ],
                "summary": "Получение клиента по ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID клиента",
                        "name": "clientId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.User"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Client not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.User": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "client_id": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                }
            }
        }
    }
}